requirements.txt
requests
shortuuid
streamlink
ffmpeg-python
